Refactor: new scripts structute
