+++ Developments/PR/utils/utils.py
@@ -1,11 +1,208 @@
 __author__ = 'Mikhail Hushchyn'
 
-# This code is taken from the FairShip/python/shipPatRec.py and modified.
-
 import ROOT
 import numpy
 
-from utils import fracMCsame
+
+########################################################################################################################
+
+def decodeDetectorID(detID):
+    """
+    Decodes detector ID.
+
+    Parameters
+    ----------
+    detID : int or array-like
+        Detector ID values.
+
+    Returns
+    -------
+    statnb : int or array-like
+        Station numbers.
+    vnb : int or array-like
+        View numbers.
+    pnb : int or array-like
+        Plane numbers.
+    lnb : int or array-like
+        Layer numbers.
+    snb : int or array-like
+        Straw tube numbers.
+    """
+
+    statnb = detID // 10000000
+    vnb = (detID - statnb * 10000000) // 1000000
+    pnb = (detID - statnb * 10000000 - vnb * 1000000) // 100000
+    lnb = (detID - statnb * 10000000 - vnb * 1000000 - pnb * 100000) // 10000
+    snb = detID - statnb * 10000000 - vnb * 1000000 - pnb * 100000 - lnb * 10000 - 2000
+
+    return statnb, vnb, pnb, lnb, snb
+
+########################################################################################################################
+
+def fracMCsame(trackids):
+    """
+    Estimates max fraction of true hit labels for a recognized track.
+    trackids : array_like
+        hit indexes of a recognized track
+
+    Retunrs
+    -------
+    frac : float
+        Max fraction of true hit labels.
+    tmax : int
+        True hit label with max fraction in a recognized track.
+    """
+
+    track={}
+    nh=len(trackids)
+    for tid in trackids:
+        if tid==999:
+            nh-=1
+            continue
+        if track.has_key(tid):
+            track[tid]+=1
+        else:
+            track[tid]=1
+    #now get track with largest number of hits
+    if track != {}:
+        tmax=max(track, key=track.get)
+    else:
+        track = {-999:0}
+        tmax = -999
+
+    frac=0.
+    if nh>0: frac=float(track[tmax])/float(nh)
+
+    return frac,tmax
+
+########################################################################################################################
+
+def Digitization(sTree, SmearedHits):
+    """
+    Digitizes hit for the track pattern recognition.
+
+    Parameters
+    ----------
+    sTree : root file
+        Events in raw format.
+    SmearedHits : list of dicts
+        List of smeared hits. A smeared hit is a dictionary:
+        {'digiHit':key,'xtop':top x,'ytop':top y,'z':top z,'xbot':bot x,'ybot':bot y,'dist':smeared dist2wire}
+
+    Retruns
+    -------
+    X : ndarray-like
+        Information about active straw tubes: [[xtop, ytop, ztop, xbot, ybot, zboy, dist2wire, detID], [...], ...]
+    """
+
+    Hits = []
+
+    for i in range(len(SmearedHits)):
+        xtop=SmearedHits[i]['xtop']
+        xbot=SmearedHits[i]['xbot']
+        ytop=SmearedHits[i]['ytop']
+        ybot=SmearedHits[i]['ybot']
+        ztop=SmearedHits[i]['z']
+        zbot=SmearedHits[i]['z']
+        distance=SmearedHits[i]['dist']
+        detid=sTree.strawtubesPoint[i].GetDetectorID()
+
+        ahit=[xtop, ytop, ztop, xbot, ybot, zbot, float(distance), int(detid)]
+        Hits.append(ahit)
+
+    return numpy.array(Hits)
+
+########################################################################################################################
+
+def smearHits(sTree, ShipGeo, modules, no_amb=None):
+    """
+    Smears hits.
+
+    Parameters
+    ----------
+    sTree : root file
+        Events in raw format.
+    ShipGeo : object
+        Contains SHiP detector geometry.
+    modules : object
+        Contains SHiP detector geometry.
+    no_amb : boolean
+        If False - no hit smearing.
+
+    Returns
+    -------
+    SmearedHits : list of dicts
+        List of smeared hits. A smeared hit is a dictionary:
+        {'digiHit':key,'xtop':top x,'ytop':top y,'z':top z,'xbot':bot x,'ybot':bot y,'dist':smeared dist2wire}
+    """
+
+    random = ROOT.TRandom()
+    ROOT.gRandom.SetSeed(13)
+
+
+    # Loking for duplicates
+    hitstraws = {}
+    duplicatestrawhit = []
+
+    nHits = sTree.strawtubesPoint.GetEntriesFast()
+    for i in range(nHits):
+
+        ahit = sTree.strawtubesPoint[i]
+
+        if (str(ahit.GetDetectorID())[:1]=="5") :
+            continue
+
+        strawname=str(ahit.GetDetectorID())
+
+        if hitstraws.has_key(strawname):
+
+            if ahit.GetX()>hitstraws[strawname][1]:
+                duplicatestrawhit.append(i)
+            else:
+                duplicatestrawhit.append(hitstraws[strawname][0])
+                hitstraws[strawname]=[i,ahit.GetX()]
+        else:
+
+            hitstraws[strawname]=[i,ahit.GetX()]
+
+
+    # Smear strawtube points
+    SmearedHits = []
+    key = -1
+
+    for i in range(nHits):
+
+        ahit = sTree.strawtubesPoint[i]
+
+        # if i in duplicatestrawhit:
+        #     continue
+        #
+        # if (str(ahit.GetDetectorID())[:1]=="5") :
+        #     continue
+        #
+        # if (((ahit.GetX()/245.)**2 + (ahit.GetY()/495.)**2) >= 1.):
+        #     continue
+
+        key+=1
+        detID = ahit.GetDetectorID()
+        top = ROOT.TVector3()
+        bot = ROOT.TVector3()
+
+        modules["Strawtubes"].StrawEndPoints(detID,bot,top)
+
+        #distance to wire, and smear it.
+        dw  = ahit.dist2Wire()
+        smear = dw
+        if not no_amb:
+            random = ROOT.TRandom()
+            smear = abs(random.Gaus(dw,ShipGeo.strawtubes.sigma_spatial))
+
+        SmearedHits.append( {'digiHit':key,'xtop':top.x(),'ytop':top.y(),'z':top.z(),'xbot':bot.x(),'ybot':bot.y(),'dist':smear} )
+
+
+    return SmearedHits
+
+########################################################################################################################
 
 def get_track_ids(stree, smeared_hits):
     """
@@ -34,6 +231,7 @@ def get_track_ids(stree, smeared_hits):
 
     return numpy.array(y)
 
+########################################################################################################################
 
 def get_fitted_trackids(y, reco_tracks):
     """
@@ -72,6 +270,7 @@ def get_fitted_trackids(y, reco_tracks):
 
     return fittedtrackids, fittedtrackfrac
 
+########################################################################################################################
 
 def getReconstructibleTracks(iEvent,sTree,sGeo, reconstructiblerequired, threeprong, ShipGeo):
     """
@@ -355,3 +554,5 @@ def getReconstructibleTracks(iEvent,sTree,sGeo, reconstructiblerequired, threepr
         pass
     if debug==1: print "Tracks with required HNL decay particles",MCTrackIDs
     return MCTrackIDs
+
+########################################################################################################################
