--- Developments/track_pattern_recognition/requirements.txt
@@ -1,3 +1,0 @@
-numpy >= 1.12.1
-scipy >= 0.19.0
-scikit-learn >= 0.17.1
